FEAT: Allow arbitrary source location
The Cloudformation templates are tuned to use GitHub repositories, but
this means that to test changes a new commit must be made.

Simplify the process of calculating the --path-prefix parameter, so any
arbitrary path can be easily used. This means that, for example, an s3
bucket can be used, although the https protocol must be used and the
objects will need to be publicly accessible.